# laravel.vpd121.com/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api Project",
        "description": "Demo my Project ",
        "contact": {
            "email": "admin@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/category": {
            "get": {
                "tags": [
                    "Category"
                ],
                "operationId": "327644f4d836240fa297b2822ab6ca92",
                "responses": {
                    "200": {
                        "description": "List Categories."
                    }
                }
            },
            "post": {
                "tags": [
                    "Category"
                ],
                "operationId": "75c388e2c6bd800c115e7884d915a82f",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [],
                                "properties": {
                                    "image": {
                                        "type": "string"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Add Category."
                    }
                }
            }
        },
        "/api/category/{id}": {
            "get": {
                "tags": [
                    "Category"
                ],
                "operationId": "b5e0594038139a2979e671987221c993",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Ідентифікатор категорії",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List Categories."
                    },
                    "404": {
                        "description": "Wrong id",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, wrong Category Id has been sent. Pls try another one."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Category"
                ],
                "operationId": "d2437e8e0e200592efe7305af1c302d7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Ідентифікатор категорії",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успішне видалення категорії"
                    },
                    "404": {
                        "description": "Категорії не знайдено"
                    },
                    "401": {
                        "description": "Не авторизований"
                    }
                }
            }
        },
        "/api/category/edit/{id}": {
            "post": {
                "tags": [
                    "Category"
                ],
                "operationId": "fa02b8a4748a6e7b435d982b5d8d9db4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Ідентифікатор категорії",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "image": {
                                        "type": "string"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Add Category."
                    }
                }
            }
        },
        "/api/product": {
            "get": {
                "tags": [
                    "Product"
                ],
                "operationId": "5cc57bec35d0f2045983cc8dcd00eccd",
                "responses": {
                    "200": {
                        "description": "Get All Products."
                    }
                }
            },
            "post": {
                "tags": [
                    "Product"
                ],
                "operationId": "a2a748b2864d35d1e50976b938ba8cd9",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [],
                                "properties": {
                                    "price": {
                                        "type": "number"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "category_id": {
                                        "type": "number"
                                    },
                                    "description": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Add Product."
                    }
                }
            }
        },
        "/api/product/{id}": {
            "get": {
                "tags": [
                    "Product"
                ],
                "operationId": "0847d7d2b6de240f2532a6a0a28060a8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Ідентифікатор продукту",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List Products."
                    },
                    "404": {
                        "description": "Wrong id",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, wrong Product Id has been sent. Pls try another one."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Product"
                ],
                "operationId": "118d6027a434cda88b95d382d657bd05",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Ідентифікатор продукту",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успішне видалення продукту"
                    },
                    "404": {
                        "description": "Продукт не знайдено"
                    },
                    "401": {
                        "description": "Не авторизований"
                    }
                }
            }
        },
        "/api/product/edit/{id}": {
            "post": {
                "tags": [
                    "Product"
                ],
                "operationId": "136e767b610e5949320b0813a391424a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Ідентифікатор продукту",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "category_id": {
                                        "type": "number"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "price": {
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Update Product"
                    },
                    "404": {
                        "description": "Product not found"
                    }
                }
            }
        },
        "/api/product_images": {
            "get": {
                "tags": [
                    "ProductImage"
                ],
                "operationId": "83e08432ef5eff3f18873b1c157149d4",
                "responses": {
                    "200": {
                        "description": "Get All Product Images."
                    }
                }
            },
            "post": {
                "tags": [
                    "ProductImage"
                ],
                "operationId": "845179ad1d15d9f07c272003192c426a",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [],
                                "properties": {
                                    "product_id": {
                                        "type": "number"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "priority": {
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Add Product."
                    }
                }
            }
        },
        "/api/product_images/{id}": {
            "get": {
                "tags": [
                    "ProductImage"
                ],
                "operationId": "17b9ebda9cf49761a5af0e649022a083",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Ідентифікатор картинки",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List Pictures."
                    },
                    "404": {
                        "description": "Wrong id",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, wrong Picture Id has been sent. Pls try another one."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "ProductImage"
                ],
                "operationId": "d8183c0fc9f5cf86069d2d42d2a6f54e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Ідентифікатор Картинки",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успішне видалення картинки"
                    },
                    "404": {
                        "description": "Картинку не знайдено"
                    },
                    "401": {
                        "description": "Не авторизований"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}